Fix Issues: change everything to size_t, add structure to esc sequences function and improved readability

+++ inc/cetushell.h
@@ -32,8 +32,8 @@ typedef struct		s_buff
 
 typedef struct		s_point
 {
-	int				x;
-	int				y;
+	size_t				x;
+	size_t				y;
 }					t_point;
 
 typedef struct		s_cursor
